FCN-342: fixtures plus storybook ai slop fix
